Dirty hack to load GEOS from C65 mode directly
Basically it just "prefixes" the puc'ed binary with a C65 basic stub to
set-up some C64-compatible settings in relation of memory and relocting
stuff followed by transferring control to the old un-puc' stuff.

In this way, GEOS65 can be loaded from C65 mode as well (no need for
GO64 first). The result works in my Xemu/C65 emulator, however GEOS
won't work, as it reads CIA-2 reg 0, maybe because of IEC serial bus.
However that is not emulated, and a new disk driver is needed for F011
FDC in Commodore 65. There is *some* driver for that already, but it's
not clear it's for Mega65 only or C65 too. Moreover, the current GEOS
repository of mist64 does not allow to include custom disk driver other
than 1541, at least I couldn't find it out, how to do so.